requirements.txt
annotated-types==0.7.0
greenlet==3.4.0
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pdf2image==1.17.0
pillow==12.2.0
playwright==1.58.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
RapidFuzz==3.14.5
six==1.17.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
